Skeleton code for the index entry addition is added to application and fixed a bug in which the follower component were not resetting the election phase correctly. In addition to this, couple of minor fixes.

+++ leader-election/election-core/src/main/java/se/sics/p2ptoolbox/election/core/util/ElectionHelper.java
@@ -48,7 +48,7 @@ public class ElectionHelper {
 
         oldAddressSet.retainAll(currentAddressSet);
 
-        return ((oldSize == newSize) && oldAddressSet.size() >= convergenceFactor * currentAddressSet.size());
+        return ((oldSize == newSize) && oldAddressSet.size() >=  (int)(convergenceFactor * currentAddressSet.size()));
     }
